https://skia.googlesource.com/skia.git/+log/3633597a4fa7..cc360436396e

2023-07-14 johnstiles@google.com Another fix for Clang format-specifier warning.
2023-07-14 skia-autoroll@skia-public.iam.gserviceaccount.com Roll vulkan-deps from e1b8f324086e to fcbe6bbcf4a8 (4 revisions)
2023-07-14 robertphillips@google.com Roll Vulkan Memory Allocator to v3.0.1
2023-07-14 johnstiles@google.com Fix format specifier for backend enum.

If this roll has caused a breakage, revert this CL and stop the roller
using the controls here:
https://skia-autoroll.corp.goog/r/android-master-autoroll
Please CC djsollen@google.com,nscobie@google.com,rmistry@google.com on the revert to ensure that a human
is aware of the problem.

To report a problem with the AutoRoller itself, please file a bug:
https://bugs.chromium.org/p/skia/issues/entry?template=Autoroller+Bug

Documentation for the AutoRoller is here:
https://skia.googlesource.com/buildbot/+doc/main/autoroll/README.md

Test: Presubmit checks will test this change.
Exempt-From-Owner-Approval: The autoroll bot does not require owner approval.
Bug: b/230029570
Bug: b/186777432

+++ src/gpu/ganesh/GrAHardwareBufferUtils.cpp
@@ -348,9 +348,9 @@ static GrBackendTexture make_vk_backend_texture(
 
     VkFormat format;
     if (!backendFormat.asVkFormat(&format)) {
-        SkDebugf("asVkFormat failed (valid: %d, backend: %d)",
+        SkDebugf("asVkFormat failed (valid: %d, backend: %u)",
                  backendFormat.isValid(),
-                 backendFormat.backend());
+                 (unsigned)backendFormat.backend());
         return GrBackendTexture();
     }
